Fix File struction
